lambda/embeddings/requirements.txt
boto3
botocore
psycopg2-binary
numpy
Pillow
requests
sentence-transformers
transformers
